Elementor\Editor moved to Elementor\Core\Editor\Editor
